(refactor): move manual tests into integration test dir
- move build-withConfig into integration-tests/fixtures/
  - move its package.json deps into root devDeps
    - it's much easier to work with this way vs. as subdirectories
      with separate package.json and node_modules
      - and consistent with how the E2E tests work too
    - (deps): remove eslint-config-postcss as it wasn't used anyway

- move fixtures/ to test/fixtures as they're specific to E2E tests

- move test/jest.config.json to root jest.config.js
  - so it's used by default, no need for --config
- significantly simplify jest.config.js now that it's in root too
  - and the simplification allows it to be used for both test dirs

(docs): update fixtures README to remove "manual" references
(docs): reword all mentions of tests to say E2E tests
- that's what they currently are, though I'd like to refactor many to
  be unit tests
(docs): add tests README and integration test fixtures README
