+++ package.json
@@ -26,7 +26,7 @@
     "lint": "yarn build && yarn lint:post-build",
     "lint:post-build": "node dist/index.js lint src test --ignore-pattern 'test/tests/lint'",
     "test": "yarn build && yarn test:post-build",
-    "test:post-build": "node dist/index.js test --config ./test/jest.config.json",
+    "test:post-build": "node dist/index.js test",
     "start": "tsc -p tsconfig.json --watch",
     "release": "np"
   },
@@ -112,13 +112,18 @@
     "@types/rollup-plugin-sourcemaps": "^0.4.2",
     "@types/sade": "^1.6.0",
     "@types/semver": "^7.1.0",
+    "autoprefixer": "^9.7.4",
+    "cssnano": "^4.1.10",
     "doctoc": "^1.4.0",
     "husky": "^4.2.2",
     "np": "^6.2.0",
     "pretty-quick": "^2.0.0",
     "ps-tree": "^1.2.0",
     "react": "^16.8.6",
-    "semver": "^7.1.1"
+    "rollup-plugin-postcss": "^2.5.0",
+    "semver": "^7.1.1",
+    "tiny-invariant": "^1.1.0",
+    "tiny-warning": "^1.0.3"
   },
   "husky": {
     "hooks": {
