+++ test/README.md
@@ -1,0 +1,6 @@
+# Tests
+
+There are two main directories here:
+
+- `tests` contains end-to-end (E2E) tests of the CLI
+- `integration-tests` contains tests ensuring that common or recommended plugins work properly together with TSDX
