--- test/fixtures/build-withConfig/package.json
@@ -1,17 +1,0 @@
-{
-  "scripts": {
-    "build": "tsdx build --extractErrors"
-  },
-  "name": "build-withconfig",
-  "license": "MIT",
-  "dependencies": {
-    "tiny-invariant": "^1.0.6",
-    "tiny-warning": "^1.0.3"
-  },
-  "devDependencies": {
-    "autoprefixer": "^9.6.1",
-    "cssnano": "^4.1.10",
-    "eslint-config-postcss": "^3.0.7",
-    "rollup-plugin-postcss": "^2.0.3"
-  }
-}
