+++ test/integration-tests/fixtures/README.md
@@ -1,0 +1,3 @@
+# Integration Test Fixtures Directory
+
+- `build-withConfig` lets us check that `tsdx.config.js` works as expected
