+++ test/integration-tests/fixtures/build-withConfig/errors/ErrorDev.js
