+++ test/integration-tests/fixtures/build-withConfig/src/foo.ts
