+++ test/integration-tests/fixtures/build-withConfig/src/index.ts
