+++ test/integration-tests/fixtures/build-withConfig/tsconfig.json
