+++ test/integration-tests/fixtures/build-withConfig/tsdx.config.js
