--- test/jest.config.json
@@ -1,12 +1,0 @@
-{
-  "testEnvironment": "node",
-  "roots": ["<rootDir>/tests"],
-  "collectCoverageFrom": ["**/*.js"],
-  "transform": {
-    "^.+\\.ts?$": "ts-jest"
-  },
-  "testMatch": [
-    "<rootDir>/tests/**/__tests__/**/*.[jt]s?(x)",
-    "<rootDir>/tests/**/*(*.)@(spec|test).[tj]s?(x)"
-  ]
-}
