+++ test/tests/fixtures/build-default/package2.json
