+++ test/tests/fixtures/build-default/src/foo.ts
