+++ test/tests/fixtures/build-default/src/index.ts
