+++ test/tests/fixtures/build-default/src/syntax/jsx-import/JSX-B.jsx
