+++ test/tests/fixtures/build-default/src/syntax/jsx-import/JSX-import-JSX.jsx
