+++ test/tests/fixtures/build-default/src/syntax/optional-chaining.ts
