+++ test/tests/fixtures/build-invalid/package.json
