+++ test/tests/fixtures/build-invalid/src/index.ts
