+++ test/tests/fixtures/build-invalid/tsconfig.json
