+++ test/tests/fixtures/build-withTsconfig/package.json
