+++ test/tests/fixtures/build-withTsconfig/src/index.ts
