+++ test/tests/fixtures/build-withTsconfig/src/tsconfig.json
