+++ test/tests/fixtures/build-withTsconfig/tsconfig.base.json
