+++ test/tests/fixtures/build-withTsconfig/tsconfig.json
