+++ test/tests/tsdx-build-default.test.js
@@ -1,16 +1,17 @@
 const shell = require('shelljs');
-const util = require('../fixtures/util');
+const util = require('../utils/fixture');
 const { execWithCache } = require('../utils/shell');
 
 shell.config.silent = false;
 
+const testDir = 'tests';
 const fixtureName = 'build-default';
 const stageName = `stage-${fixtureName}`;
 
 describe('tsdx build :: zero-config defaults', () => {
   beforeAll(() => {
     util.teardownStage(stageName);
-    util.setupStageWithFixture(stageName, fixtureName);
+    util.setupStageWithFixture(testDir, stageName, fixtureName);
   });
 
   it('should compile files into a dist directory', () => {
